// web/packages/agenta-playground-ui/src/components/InputMappingModal/types.ts
export interface MappingStatusInfo {
    color: "red" | "orange" | "blue" | "green" | "default"
    label: string
    icon: React.ReactNode
}